moved the event handler to a diferent file
