+++ src/modules/service-provider/service-provider.module.ts
@@ -6,10 +6,17 @@ import { ServiceProviderFactory } from './domain/service-provider.factory.js';
 import { KeycloakUserService } from '../keycloak-administration/index.js';
 import { KeycloakAdministrationModule } from '../keycloak-administration/keycloak-administration.module.js';
 import { KeycloakAdministrationService } from '../keycloak-administration/domain/keycloak-admin-client.service.js';
+import { CreateGroupAndRoleHandler } from './repo/service-provider-event-handler.js';
 
 @Module({
     imports: [LoggerModule.register(ServiceProviderModule.name), KeycloakAdministrationModule],
-    providers: [ServiceProviderRepo, ServiceProviderFactory, KeycloakUserService, KeycloakAdministrationService],
-    exports: [ServiceProviderRepo, ServiceProviderFactory, KeycloakUserService],
+    providers: [
+        ServiceProviderRepo,
+        ServiceProviderFactory,
+        KeycloakUserService,
+        KeycloakAdministrationService,
+        CreateGroupAndRoleHandler,
+    ],
+    exports: [ServiceProviderRepo, ServiceProviderFactory, KeycloakUserService, CreateGroupAndRoleHandler],
 })
 export class ServiceProviderModule {}
